Koluvu-backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
channels
django-cors-headers
python-dotenv
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
django-filter==24.3
# CAPTCHA
django-recaptcha==4.0.0
# Security
django-ratelimit==4.1.0
django-otp==1.5.4
cryptography==42.0.8
# ATS System Dependencies
PyPDF2==3.0.1
python-docx==1.2.0
setuptools
nltk==3.9.1
scikit-learn==1.7.0
Pillow
numpy==1.26.4